requirements.txt
torch~=1.12.1
torchvision~=0.13.1
ftfy
regex~=2021.11.10
tqdm~=4.64.0
opencv-python~=4.6.0.66
boto3
requests~=2.27.1
pandas~=1.3.4
pytube~=12.1.2
numpy~=1.22.3
argparse~=1.4.0
psutil~=5.8.0
Pillow~=9.1.0
loguru~=0.6.0
setuptools~=65.5.0
clip~=0.2.0
moviepy~=1.0.3
matplotlib
scikit-learn
jupyter